\name{exampleData}
\alias{exampleData}
\docType{data}
\title{
   Description of the data in the package
}
\description{
   \code{chromLength} is the length of 24 human chromosome. The first column is chromosome name and 
   the second one is the length.
   \code{genelocate} is location of all genes in the human chromosome. The first column is gene symbol, 
   second is the chromosome, third is the start position and fourth is the end position.
   \code{sourceOmics} is the copy number alteration data of human colorectal cancer, which was downloaded 
   from firehose (http://gdac.broadinstitute.org) and contains 40 samples and genes from chromosome 20.
   \code{targetOmics} is the RNASeq data of human colorectal cancer, which was dowloaded from the Nature
   paper (Proteogenomic characterization of human colon and rectal cancer. Nature. 2014 
   Sep 18;513(7518):382-7. (PMID: 25043054)) and contains 3764 genes and 40 samples.
}

\keyword{datasets}
