\name{multiOmicsViz-package}
\alias{multiOmicsViz-package}

\docType{package}
\title{
	Plot the effect of one omics data on other omics data along the chromosome
}
\description{
	Calculate the spearman correlation between the source omics data and other 
	target omics data, identify the significant correlations and plot the significant 
	correlations on the heat map in which the x-axis and y-axis are ordered by the 
	chromosomal location.
}
\details{
	\tabular{ll}{
		Package: \tab multiOmicsViz\cr
		Type: \tab Package\cr
		License: \tab LGPL\cr
		LazyLoad: \tab yes\cr
	}
}

\author{
	Jing Wang
	Maintainer: Jing Wang <jingwang.uestc@gmail.com>
}

\references{
	Proteogenomic characterization of human colon and rectal cancer. Nature. 
	2014 Sep 18;513(7518):382-7. (PMID: 25043054)
}

\seealso{
	\code{\link{multiOmicsViz}}
}

\keyword{ package }
